navigation 정리. 파일 분리. 폴더 정리. 버전 업데이트. curating, archive 구현해야함. api 연결해야함.
