+++ frontend/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Wed May 28 14:33:33 KST 2025
-gradle.version=8.11.1
+#Mon Jun 30 15:23:34 KST 2025
+gradle.version=8.13
