--- frontend/app/src/main/java/com/stellan/challang/ui/navigation/HomeNavHost.kt
@@ -1,19 +1,0 @@
-package com.stellan.challang.ui.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
-import com.stellan.challang.ui.screen.ArchiveScreen
-import com.stellan.challang.ui.screen.CuratingScreen
-import com.stellan.challang.ui.screen.MyPageScreen
-
-@Composable
-fun HomeNavHost(
-    tab: String,
-    parentNavController: NavController
-) {
-    when (tab) {
-        "home" -> CuratingScreen()
-        "mypage" -> MyPageScreen(parentNavController)
-        "archive" -> ArchiveScreen()
-    }
-}
