+++ frontend/app/src/main/java/com/stellan/challang/ui/screen/mypage/PrivacyPolicyScreen.kt
@@ -1,0 +1,83 @@
+package com.stellan.challang.ui.screen.mypage
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.stellan.challang.ui.theme.PaperlogyFamily
+import com.stellan.challang.ui.util.loadTextFromAsset
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun PrivacyPolicyScreen(onBack: () -> Unit) {
+    val context = LocalContext.current
+    val content = loadTextFromAsset(context, "privacy_policy.txt")
+
+    Scaffold(
+        containerColor = Color(0xFFDDF0F0),
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(
+                        text = "개인정보처리방침",
+                        fontFamily = PaperlogyFamily,
+                        fontSize = 20.sp,
+                        fontWeight = FontWeight.Bold
+                    )
+                },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(
+                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                            contentDescription = "뒤로가기"
+                        )
+                    }
+                },
+                colors = TopAppBarDefaults.topAppBarColors(
+                    containerColor = Color(0xFFDDF0F0),
+                    titleContentColor = Color.Black,
+                    navigationIconContentColor = Color.Gray
+                )
+            )
+        }
+    ) { padding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(padding)
+                .padding(horizontal = 20.dp, vertical = 12.dp)
+        ) {
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .verticalScroll(rememberScrollState())
+            ) {
+                Text(
+                    text = content,
+                    fontFamily = PaperlogyFamily,
+                    fontSize = 14.sp,
+                    lineHeight = 20.sp,
+                    color = Color.Black
+                )
+            }
+        }
+    }
+}
